# src/coordinate_transformation_api/assets/openapi.yaml
openapi: 3.1.0
info:
  title: NSGI Coordinate Transformation API
  version: '' # replaced in code
  description: >
    The Coordinate Transformation API performs coordinate transformations on GeoJSON and CityJSON files using the official RDNAPTRANS™ transformation procedure and other transformations defined or recommended by NSGI.

  contact:
    name: NSGI
    email: rd@kadaster.nl
    url: https://www.nsgi.nl
  termsOfService: https://creativecommons.org/licenses/by/4.0/
  x-repositoryUrl: https://github.com/GeodetischeInfrastructuur/coordinate-transformation-api
  x-keywords:
    - geospatial
    - api
    - transformations
    - coordinates
    - GeoJSON
    - CityJSON
    - RDNAPTRANS
servers:
  - url: http://example.org # replaced in code
...
paths:
...
  /openapi:
    get:
      operationId: get-oas
      tags:
        - About
      summary: OAS
      description: The Open API Specification (OAS) document of this API.
      responses:
        '200':
          $ref: '#/components/responses/200-oas'
        '500':
          $ref: '#/components/responses/500'
  /conformance:
    get:
      operationId: get-conformance
      tags:
        - About
      summary: Conformance page
      description: >
        OGC API Common Conformance page describing the capabilities of this API. Best effort has been made to comply to the OGC API Commons standard, but due to the nature of this API, full compliance is not possible. See for details the [docs](https://github.com/GeodetischeInfrastructuur/coordinate-transformation-api/blob/main/docs/OGC-API-Common.md).

      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConformanceDeclaration'
            text/html:
              schema:
                type: string
          headers:
            api-version:
              $ref: '#/components/headers/api-version'
        '500':
          $ref: '#/components/responses/500'
  /crss:
    get:
      operationId: get-crss
      tags:
        - CRS
      summary: Get all CRSs supported by the transformation API
      description: Returns definitions of all supported coordinate reference systems (CRSs)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Crs'
                type: array
                title: CRS List
        '500':
          $ref: '#/components/responses/500'
  /crss/{crs-id}:
    get:
      operationId: get-crs
      tags:
        - CRS
      summary: Get a single CRS
      description: >
        Returns the definition of a single coordinate reference system (CRS) by `crs-id` (format: `{authority:identifier}`, for example: `EPSG:28992`).

      parameters:
        - name: crs-id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/CrsEnum'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crs'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404Crs'
        '500':
          $ref: '#/components/responses/500'
  /check-density:
    post:
      operationId: post-density-check
      tags:
        - Check-density
      summary: >
        Check segment length of (multi)polygon and (multi)linestring geometries based on the geodesic (ellipsoidal great-circle) distance on the GRS80 ellipsoid

      description: |
        Check segment length of polygon and linestring geometries. Returns a `DensityCheckReport` containing a FeatureCollection containing all line segments exceeding the maximum segment length 'maxSegmentLength' or maximum segment deviation `maxSegmentDeviation`.
      parameters:
        - $ref: '#/components/parameters/sourceCrs'
        - $ref: '#/components/parameters/contentCrs'
        - $ref: '#/components/parameters/maxSegmentLength'
        - $ref: '#/components/parameters/maxSegmentDeviation'
      requestBody:
        required: true
        description: |
          One of:
          - GeoJSON Feature
          - GeoJSON FeatureCollection
          - GeoJSON Geometry
          - GeoJSON GeometryCollection

          GeoJSON FeatureCollection objects may contain a `.crs` member containing a named CRS, to indicate the source CRS. See the `GeoJsonCrs` schema for more details.

          > **Note:** The `crs` member is not part of the [FeatureCollection schema](schemas/geojson/FeatureCollection.json).

          When `crs` is specified in the request body, the `source-crs` query parameter or `content-crs` request header may be omitted. Order of precedence of CRS input is: request body > query parameters > headers.
        content:
          application/json:
            schema:
              oneOf:
                - $ref: schemas/geojson/Feature.json
                - $ref: schemas/geojson/FeatureCollection.json
                - $ref: schemas/geojson/Geometry.json
                - $ref: schemas/geojson/GeometryCollection.json
              example:
                type: FeatureCollection
                name: lijnen
                crs:
                  type: name
                  properties:
                    name: urn:ogc:def:crs:EPSG::28992
                features:
                  - type: Feature
                    properties: {}
                    geometry:
                      type: LineString
                      coordinates:
                        - - 156264.9063
                          - 601302.5889
                        - - 165681.9644
                          - 605544.3131
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DensityCheckReport'
          headers:
            api-version:
              $ref: '#/components/headers/api-version'
            content-crs:
              description: Coordinate reference system (CRS) of the input coordinates
              $ref: '#/components/headers/content-crs'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /densify:
    post:
      operationId: post-densify
      tags:
        - Experimental
      summary: >
        Densify (multi)polygon and (multi)linestring geometries using the geodesic (ellipsoidal great-circle) trajectory on the GR80 ellipsoid

      description: |
        **Warning**: This endpoint is currently in an experimental phase!

        A POST endpoint that accepts a GeoJSON object and densifies the
        geometries using the maximum segment length 'maxSegmentLength' or
        maximum segment deviation `maxSegmentDeviation` threshold.
      parameters:
        - $ref: '#/components/parameters/sourceCrs'
        - $ref: '#/components/parameters/contentCrs'
        - $ref: '#/components/parameters/maxSegmentLength'
        - $ref: '#/components/parameters/maxSegmentDeviation'
      requestBody:
        required: true
        description: |
          One of:
          - GeoJSON Feature
          - GeoJSON FeatureCollection
          - GeoJSON Geometry
          - GeoJSON GeometryCollection

          GeoJSON FeatureCollection objects may contain a `.crs` member containing a named CRS, to indicate the source CRS. See the `GeoJsonCrs` schema for more details.

          > **Note:** The `crs` member is not part of the [FeatureCollection schema](schemas/geojson/FeatureCollection.json).

          When `crs` is specified in the request body, the `source-crs` query parameter or `content-crs` request header may be omitted. Order of precedence of CRS input is: request body > query parameters > headers.
        content:
          application/json:
            schema:
              oneOf:
                - $ref: schemas/geojson/Feature.json
                - $ref: schemas/geojson/FeatureCollection.json
                - $ref: schemas/geojson/Geometry.json
                - $ref: schemas/geojson/GeometryCollection.json
            example:
              type: Feature
              properties:
                id: 1
              geometry:
                type: LineString
                coordinates:
                  - - 156264.9063
                    - 601302.5889
                  - - 165681.9644
                    - 605544.3131
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: schemas/geojson/FeatureCollection.json
              example:
                type: FeatureCollection
                name: lijnen
                crs:
                  type: name
                  properties:
                    name: urn:ogc:def:crs:EPSG::28992
                features:
                  - type: Feature
                    properties: {}
                    geometry:
                      type: LineString
                      coordinates:
                        - - 156264.9063
                          - 601302.5889
                        - - 165681.9644
                          - 605544.3131
          headers:
            api-version:
              $ref: '#/components/headers/api-version'
            content-crs:
              $ref: '#/components/headers/content-crs'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
components:
...
  headers:
    api-version:
      description: Open API Specification version number
      schema:
        type: string
      example: 3.0.0
...
  responses:
    200-oas:
      description: OK
      content:
        application/openapi+json: {}
        text/html:
          schema:
            type: string
      headers:
        api-version:
          $ref: '#/components/headers/api-version'
...
    '400':
      description: Bad request. The request body does not contain a valid payload or the query is not supported by the API.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            '400':
              $ref: '#/components/examples/400Example'
      headers:
        api-version:
          $ref: '#/components/headers/api-version'
    '500':
      description: An internal server error occurred.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            '500':
              $ref: '#/components/examples/500Example'
      headers:
        api-version:
          $ref: '#/components/headers/api-version'
  schemas:
...
    ConformanceDeclaration:
      example:
        conformsTo: []
      description: |-
        The URIs of all conformance classes supported by the server.

        To support "generic" clients that want to access multiple
        OGC API Features implementations - and not "just" a specific
        API / server, the server declares the conformance
        classes it implements and conforms to.
      properties:
        conformsTo:
          items:
            type: string
          type: array
      type: object
...
    Crs:
      properties:
        authority:
          title: Authority
          type: string
        axes:
          items:
            $ref: '#/components/schemas/Axis'
          title: Axes
          type: array
        crs_auth_identifier:
          $ref: '#/components/schemas/CrsEnum'
        identifier:
          title: Identifier
          type: string
        name:
          title: Name
          type: string
        type_name:
          title: Type Name
          type: string
      required:
        - name
        - type_name
        - authority
        - identifier
        - axes
      title: CRS
      type: object
    CrsEnum:
      enum: ['EPSG:28992', 'EPSG:7415', 'EPSG:7931'] # filled in code, dummy values to suppress error by linter
      example: EPSG:28992
      type: string
...